Refactor OpenSSL DTLS into crypto mod
