+++ src/change/mod.rs
@@ -19,4 +19,5 @@ pub use sdp::{SdpAnswer, SdpApi, SdpOffer, SdpPendingOffer};
 mod direct;
 pub use direct::DirectApi;
 
-pub use crate::dtls::{DtlsCert, Fingerprint};
+pub use crate::crypto::Fingerprint;
+pub use crate::dtls::DtlsCert;
