+++ src/crypto/keying.rs
@@ -1,0 +1,24 @@
+use std::ops::Deref;
+
+/// Keying material used as master key for SRTP.
+pub struct KeyingMaterial(Vec<u8>);
+
+impl KeyingMaterial {
+    pub fn new(m: Vec<u8>) -> Self {
+        KeyingMaterial(m)
+    }
+}
+
+impl Deref for KeyingMaterial {
+    type Target = [u8];
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl std::fmt::Debug for KeyingMaterial {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "KeyingMaterial")
+    }
+}
